Codebase refactor to include, collating dataclasses, re-organising utility functions, reducing bloat in controller, removing undeed functions/parameters, general tidying up.
