+++ src/odin_pico/adapter.py
@@ -1,7 +1,6 @@
 """Adapter for a PicoScope 5444D."""
 
 import logging
-import threading
 from odin.adapters.adapter import (
     ApiAdapter,
     ApiAdapterResponse,
@@ -19,15 +18,13 @@ class PicoAdapter(ApiAdapter):
         """Initialise the PicoAdapter Object."""
         super(PicoAdapter, self).__init__(**kwargs)
 
-        self.lock = threading.Lock()
         update_loop = True
         data_output_path = self.options.get("data_output_path", "/tmp/")
         max_caps = int(self.options.get("max_caps", 100000000))
         simulate = bool(int(self.options.get("simulate", False)))
 
         logging.debug(f"Simulate set to : {simulate}")
-
-        self.pico_controller = PicoController(self.lock, update_loop, data_output_path, max_caps, simulate)
+        self.pico_controller = PicoController(update_loop, data_output_path, max_caps, simulate)
    
     def initialize(self, adapters):
         """Initialize the adapter after it has been loaded."""
