+++ src/odin_pico/pico_controller.py
@@ -12,52 +12,37 @@ from tornado.concurrent import run_on_executor
 
 from odin_pico.analysis import PicoAnalysis
 from odin_pico.buffer_manager import BufferManager
-from odin_pico.DataClasses.device_config import DeviceConfig
-from odin_pico.DataClasses.device_status import DeviceStatus
+from odin_pico.DataClasses.pico_config import DeviceConfig
+from odin_pico.DataClasses.gpib_config import GPIBConfig
+from odin_pico.DataClasses.pico_status import DeviceStatus
+
 from odin_pico.file_writer import FileWriter
 from odin_pico.pico_device import PicoDevice
-from odin_pico.pico_util import PicoUtil
+from odin_pico.Utilities.controller_util import ControllerUtil
+from odin_pico.Utilities.pico_util import PicoUtil
+
+from odin_pico.ParameterTrees.gpib_tree import GPIBTreeBuilder
+from odin_pico.ParameterTrees.pico_tree import PicoTreeBuilder
 
 class PicoController:
     """Class which holds parameter trees and manages the PicoScope capture process."""
-
     executor = futures.ThreadPoolExecutor(max_workers=2)
 
-    def __init__(self, lock, loop, path, max_caps, simulate):
+    def __init__(self, loop, path, max_caps, simulate):
         """Initialise the PicoController Class."""
         # Threading lock and control variables
-        self.lock = lock
         self.update_loop_active = loop
-    
-        self.gpib_avail = False
-        self.gpib_control = False
-        self.tec_devices = None
-        self.selected_tec = None
         self.simulate = simulate
 
-        self.temp_setpoint = None
-        self.voltage_limit = None
-        self.current_limit = None
-
-        # Initialise variables for data collection
-        self.enable = False
-        self.caps_collected = 0
-        self.current_capture = 0
-
-        # Objects for handling configuration, data storage and representing the PicoScope 5444D
+        # Objects for handling configuration, status and utilities
         self.dev_conf = DeviceConfig()
-        self.dev_conf.file.file_path = path
-        self.channels = [
-            self.dev_conf.channel_a,
-            self.dev_conf.channel_b,
-            self.dev_conf.channel_c,
-            self.dev_conf.channel_d,
-        ]
-
-        # Initialise objects
-        self.util = PicoUtil()
+        self.gpib_config = GPIBConfig()
         self.pico_status = DeviceStatus()
-        self.buffer_manager = BufferManager(self.channels, self.dev_conf)
+        self.util = PicoUtil()
+        self.ctrl_util = ControllerUtil(self)
+
+        # Initialise objects to represent different system components
+        self.buffer_manager = BufferManager(self.dev_conf)
         self.file_writer = FileWriter(self.dev_conf, self.buffer_manager, self.pico_status)
         self.analysis = PicoAnalysis(
             self.dev_conf, self.buffer_manager, self.pico_status
@@ -65,645 +50,54 @@ class PicoController:
         self.pico = PicoDevice(max_caps, self.dev_conf, self.pico_status,
                                self.buffer_manager, self.analysis, self.file_writer)
         
+        # Initialise parameter tree to None, is built in initialize_adapters with access to other adapters
         self.param_tree = None
-        self.gpib_tree = ParameterTree({
-                    "gpib_avail": (lambda: self.gpib_avail, None),
-                    "gpib_control": (lambda: self.gpib_control, lambda value: setattr(self, 'gpib_control', value)),
-        })
+        self.dev_conf.file.file_path = path
 
-        # ParameterTrees to represent different parts of the system
-        adapter_status = ParameterTree(
-            {
-                "settings_verified": (lambda: self.pico_status.flags.verify_all, None),
-                "open_unit": (lambda: self.pico_status.open_unit, None),
-                "pico_setup_verify": (lambda: self.pico_status.pico_setup_verify, None),
-                "channel_setup_verify": (
-                    lambda: self.pico_status.channel_setup_verify,
-                    None,
-                ),
-                "channel_trigger_verify": (
-                    lambda: self.pico_status.channel_trigger_verify,
-                    None,
-                ),
-                "capture_settings_verify": (
-                    lambda: self.pico_status.capture_settings_verify,
-                    None,
-                ),
-                "file_name_verify": (
-                    lambda: True,
-                    None
-                )
-            }
-        )
-
-        self.chan_params = {}
-        for name in self.dev_conf.channel_names:
-            self.chan_params[name] = ParameterTree(
-                {
-                    "channel_id": (
-                        partial(
-                            self.get_dc_value,
-                            self.dev_conf,
-                            f"channel_{name}",
-                            "channel_id",
-                        ),
-                        None,
-                    ),
-                    "active": (
-                        partial(
-                            self.get_dc_value,
-                            self.dev_conf,
-                            f"channel_{name}",
-                            "active",
-                        ),
-                        partial(
-                            self.set_dc_chan_value,
-                            self.dev_conf,
-                            f"channel_{name}",
-                            "active",
-                        ),
-                    ),
-                    "verified": (
-                        partial(
-                            self.get_dc_value,
-                            self.dev_conf,
-                            f"channel_{name}",
-                            "verified",
-                        ),
-                        None,
-                    ),
-                    "live_view": (
-                        partial(
-                            self.get_dc_value,
-                            self.dev_conf,
-                            f"channel_{name}",
-                            "live_view",
-                        ),
-                        partial(
-                            self.set_dc_chan_value,
-                            self.dev_conf,
-                            f"channel_{name}",
-                            "live_view",
-                        ),
-                    ),
-                    "coupling": (
-                        partial(
-                            self.get_dc_value,
-                            self.dev_conf,
-                            f"channel_{name}",
-                            "coupling",
-                        ),
-                        partial(
-                            self.set_dc_chan_value,
-                            self.dev_conf,
-                            f"channel_{name}",
-                            "coupling",
-                        ),
-                    ),
-                    "range": (
-                        partial(
-                            self.get_dc_value, self.dev_conf, f"channel_{name}", "range"
-                        ),
-                        partial(
-                            self.set_dc_chan_value,
-                            self.dev_conf,
-                            f"channel_{name}",
-                            "range",
-                        ),
-                    ),
-                    "offset": (
-                        partial(
-                            self.get_dc_value,
-                            self.dev_conf,
-                            f"channel_{name}",
-                            "offset",
-                        ),
-                        partial(
-                            self.set_dc_chan_value,
-                            self.dev_conf,
-                            f"channel_{name}",
-                            "offset",
-                        ),
-                    ),
-                    "pha_active": (
-                        partial(
-                            self.get_dc_value,
-                            self.dev_conf,
-                            f"channel_{name}",
-                            "pha_active",
-                        ),
-                        partial(
-                            self.set_dc_chan_value,
-                            self.dev_conf,
-                            f"channel_{name}",
-                            "pha_active",
-                        ),
-                    ),
-                }
-            )
-
-        pico_trigger = ParameterTree(
-            {
-                "active": (
-                    lambda: self.dev_conf.trigger.active,
-                    partial(self.set_dc_value, self.dev_conf.trigger, "active"),
-                ),
-                "auto_trigger": (
-                    lambda: self.dev_conf.trigger.auto_trigger_ms,
-                    partial(
-                        self.set_dc_value, self.dev_conf.trigger, "auto_trigger_ms"
-                    ),
-                ),
-                "direction": (
-                    lambda: self.dev_conf.trigger.direction,
-                    partial(self.set_dc_value, self.dev_conf.trigger, "direction"),
-                ),
-                "delay": (
-                    lambda: self.dev_conf.trigger.delay,
-                    partial(self.set_dc_value, self.dev_conf.trigger, "delay"),
-                ),
-                "source": (
-                    lambda: self.dev_conf.trigger.source,
-                    partial(self.set_dc_value, self.dev_conf.trigger, "source"),
-                ),
-                "threshold": (
-                    lambda: self.dev_conf.trigger.threshold,
-                    partial(self.set_dc_value, self.dev_conf.trigger, "threshold"),
-                ),
-            }
-        )
-
-        pico_capture = ParameterTree(
-            {
-                "pre_trig_samples": (
-                    lambda: self.dev_conf.capture.pre_trig_samples,
-                    partial(
-                        self.set_dc_value, self.dev_conf.capture, "pre_trig_samples"
-                    ),
-                ),
-                "post_trig_samples": (
-                    lambda: self.dev_conf.capture.post_trig_samples,
-                    partial(
-                        self.set_dc_value, self.dev_conf.capture, "post_trig_samples"
-                    ),
-                ),
-                "n_captures": (
-                    lambda: self.dev_conf.capture.n_captures,
-                    partial(self.set_dc_value, self.dev_conf.capture, "n_captures"),
-                ),
-                "capture_time": (
-                    lambda: self.dev_conf.capture.capture_time,
-                    partial(self.set_dc_value, self.dev_conf.capture, "capture_time"),
-                ),
-                "capture_mode": (
-                    lambda: self.dev_conf.capture.capture_type,
-                    partial(self.set_dc_value, self.dev_conf.capture, "capture_type"),
-                ),
-                "capture_delay": (
-                    lambda: self.dev_conf.capture.capture_delay,
-                    partial(self.set_dc_value, self.dev_conf.capture, "capture_delay"),
-                ),
-                "repeat_amount": (
-                    lambda: self.dev_conf.capture.repeat_amount,
-                    partial(self.set_dc_value, self.dev_conf.capture, "repeat_amount"),
-                ),
-                "capture_repeat": (
-                    lambda: self.dev_conf.capture.capture_repeat,
-                    partial(self.set_dc_value, self.dev_conf.capture, "capture_repeat"),
-                ),
-                "max_captures": (lambda: self.pico.rec_caps, None),
-                "max_time": (lambda: self.buffer_manager.estimate_max_time(), None)
-            }
-        )
-
-        pico_mode = ParameterTree(
-            {
-                "resolution": (
-                    lambda: self.dev_conf.mode.resolution,
-                    partial(self.set_dc_value, self.dev_conf.mode, "resolution"),
-                ),
-                "timebase": (
-                    lambda: self.dev_conf.mode.timebase,
-                    partial(self.set_dc_value, self.dev_conf.mode, "timebase"),
-                ),
-                "samp_time": (lambda: self.dev_conf.mode.samp_time, None),
-            }
-        )
-
-        pico_file = ParameterTree(
-            {
-                "folder_name": (
-                    lambda: self.dev_conf.file.folder_name,
-                    partial(self.set_dc_value, self.dev_conf.file, "folder_name"),
-                ),
-                "file_name": (
-                    lambda: self.dev_conf.file.file_name,
-                    partial(self.set_dc_value, self.dev_conf.file, "file_name"),
-                ),
-                "file_path": (lambda: self.dev_conf.file.file_path, None),
-                "curr_file_name": (lambda: self.dev_conf.file.curr_file_name, None),
-                "last_write_success": (
-                    lambda: self.dev_conf.file.last_write_success,
-                    None,
-                ),
-            }
-        )
-
-        pico_pha = ParameterTree(
-            {
-                "num_bins": (
-                    lambda: self.dev_conf.pha.num_bins,
-                    partial(self.set_dc_value, self.dev_conf.pha, "num_bins"),
-                ),
-                "lower_range": (
-                    lambda: self.dev_conf.pha.lower_range,
-                    partial(self.set_dc_value, self.dev_conf.pha, "lower_range"),
-                ),
-                "upper_range": (
-                    lambda: self.dev_conf.pha.upper_range,
-                    partial(self.set_dc_value, self.dev_conf.pha, "upper_range"),
-                ),
-            }
-        )
-
-        pico_settings = ParameterTree(
-            {
-                "mode": pico_mode,
-                "channels": {
-                    name: channel for (name, channel) in self.chan_params.items()
-                },
-                "trigger": pico_trigger,
-                "capture": pico_capture,
-                "file": pico_file,
-                "pha": pico_pha,
-            }
-        )
-
-        live_view = ParameterTree(
-            {
-                "lv_active_channels": (
-                    lambda: self.buffer_manager.lv_channels_active,
-                    None,
-                ),
-                "pha_counts": (lambda: (self.buffer_manager.pha_counts.tolist()), None),
-                "capture_count": (
-                    lambda: self.dev_conf.capture_run.live_cap_comp,
-                    None,
-                ),
-                "captures_requested": (lambda: self.dev_conf.capture.n_captures, None),
-                "lv_data": (lambda: self.buffer_manager.lv_channel_arrays, None),
-                "pha_bin_edges": (lambda: (self.buffer_manager.bin_edges.tolist()), None),
-                "sweep_total" : (lambda: self.dev_conf.temp_sweep.sweep_points, None),
-                "sweep_index" : (lambda: self.dev_conf.temp_sweep.sweep_index , None),
-                "pha_active_channels": (
-                    lambda: self.buffer_manager.pha_active_channels,
-                    None,
-                ),
-                "current_tbdc_time": (lambda: self.pico.elapsed_time, None),
-                "current_capture": (lambda: self.current_capture, None),
-            }
-        )
-
-        pico_commands = ParameterTree(
-            {
-                "run_user_capture": (
-                    lambda: self.pico_status.flags.user_capture,
-                    partial(self.set_dc_value, self.pico_status.flags, "user_capture"),
-                ),
-                "clear_pha": (
-                    lambda: self.analysis.clear_pha,
-                    partial(self.set_dc_value, self.analysis, "clear_pha"),
-                )
-            }
-        )
-
-        pico_flags = ParameterTree(
-            {
-                "abort_cap": (
-                    lambda: self.pico_status.flags.abort_cap,
-                    partial(self.set_dc_value, self.pico_status.flags, "abort_cap"),
-                ),
-                "system_state": (lambda: self.pico_status.flags.system_state, None),
-            }
-        )
-
-        self.pico_param_tree = ParameterTree(
-            {
-                "status": adapter_status,
-                "commands": pico_commands,
-                "settings": pico_settings,
-                "flags": pico_flags,
-                "live_view": live_view,
-            }
-        )
-
-        self.device_tree = ParameterTree({"device": self.pico_param_tree})
-
-        # Initalise the "update_loop" if control variable passed to the Pico_Controller is True
         if self.update_loop_active:
             self.update_loop()
 
         # Set initial state of the verification system
-        self.verify_settings()
+        self.ctrl_util.verify_settings()
 
     def initialize_adapters(self, adapters):
-        """Get access to all of the other adapters. If the GPIB adapter is running, 
-        build a paramtree for a currently selected device."""
+        """Get access to all of the other adapters and build complete parameter tree."""
         try:
             self.gpib = adapters['gpib'] if adapters else None
         except:
-            pass
-
-        if self.gpib:
-            devices = self.util.iac_get(self.gpib, "devices")
-            self.tec_devices = [name for name, info in devices.items()
-                if info.get("type") == "K2510"]
-
-            self.gpib_avail = self.verify_gpib()
-
-            if self.tec_devices:
-                self.selected_tec = self.tec_devices[0]
-
-                self.gpib_info = ParameterTree({
-                    "tec_setpoint": (lambda: (self.util.iac_get(self.gpib, 
-                                        ("devices/"+self.selected_tec+"/info/tec_setpoint"))), None),
-                    "tec_volt_lim": (lambda: (self.util.iac_get(self.gpib, 
-                                        ("devices/"+self.selected_tec+"/info/tec_volt_lim"))), None),
-                    "tec_curr_lim": (lambda: (self.util.iac_get(self.gpib, 
-                                        ("devices/"+self.selected_tec+"/info/tec_curr_lim"))), None),
-                    "tec_current": (lambda: (self.util.iac_get(self.gpib, 
-                                        ("devices/"+self.selected_tec+"/info/tec_current"))), None),
-                    "tec_voltage": (lambda: (self.util.iac_get(self.gpib, 
-                                        ("devices/"+self.selected_tec+"/info/tec_voltage"))), None),
-                    "tec_power": (lambda: (self.util.iac_get(self.gpib, 
-                                        ("devices/"+self.selected_tec+"/info/tec_power"))), None),
-                    "tec_temp_meas": (lambda: (self.util.iac_get(self.gpib, 
-                                        ("devices/"+self.selected_tec+"/info/tec_temp_meas"))), None),
-                })
-
-                self.gpib_set = ParameterTree({
-                    "temp_target": (
-                        lambda: self.dev_conf.temp_sweep.temp_target,
-                        lambda v: setattr(self.dev_conf.temp_sweep, "temp_target", float(v))
-                    ),
-                    "set_temp": (
-                        lambda: None,
-                        self.set_temp_single_shot
-                    ),
-                    "c_lim": (lambda: (self.util.iac_get(self.gpib, 
-                                ("devices/"+self.selected_tec+"/set/c_lim_set"))), 
-                            lambda value: self.util.iac_set(self.gpib, 
-                                ("devices/"+self.selected_tec+"/set/c_lim_set", float(value)))),
-                    "v_lim": (lambda: (self.util.iac_get(self.gpib, 
-                                ("devices/"+self.selected_tec+"/set/v_lim_set"))), 
-                            lambda value: self.util.iac_set(self.gpib, 
-                                ("devices/"+self.selected_tec+"/set/v_lim_set", float(value))))
-                })
-
-                self.gpib_temp_sweep = ParameterTree({
-                    "active" : (lambda: self.dev_conf.temp_sweep.active,
-                                lambda v: self.set_dc_value(self.dev_conf.temp_sweep, "active",  v)),
-                    "t_start": (lambda: self.dev_conf.temp_sweep.t_start,
-                                lambda v: self.set_dc_value(self.dev_conf.temp_sweep, "t_start", v)),
-                    "t_end"  : (lambda: self.dev_conf.temp_sweep.t_end,
-                                lambda v: self.set_dc_value(self.dev_conf.temp_sweep, "t_end",   v)),
-                    "t_step" : (lambda: self.dev_conf.temp_sweep.t_step,
-                                lambda v: self.set_dc_value(self.dev_conf.temp_sweep, "t_step",  v)),
-                    "tol"    : (lambda: self.dev_conf.temp_sweep.tol,
-                                lambda v: self.set_dc_value(self.dev_conf.temp_sweep, "tol",     v)),
-                    "poll_s" : (lambda: self.dev_conf.temp_sweep.poll_s,
-                                lambda v: self.set_dc_value(self.dev_conf.temp_sweep, "poll_s",  v)),
-                })
-                
-                self.gpib_tree = ParameterTree(
-                {
-                    "gpib_avail": (lambda: self.gpib_avail, None),
-                    "gpib_control": (lambda: self.gpib_control, lambda value: setattr(self, 'gpib_control', value)),
-                    "available_tecs": (lambda: self.tec_devices, None),
-                    "selected_tec": (lambda: self.selected_tec, lambda value: setattr(self, 'selected_tec', value)),
-                    "device_control_state": (lambda: (self.util.iac_get(self.gpib, 
-                                                ("devices/"+self.selected_tec+"/device_control_state"))), 
-                                            lambda value: self.util.iac_set(self.gpib, 
-                                                ("devices/"+self.selected_tec+"/"), {"device_control_state": value})),
-                    "output_state": (lambda: (self.util.iac_get(self.gpib, 
-                                        ("devices/"+self.selected_tec+"/output_state"))), 
-                                    lambda value: self.util.iac_set(self.gpib, 
-                                        ("devices/"+self.selected_tec+"/"), {"output_state": value})),
-                    "temp_over_state": (lambda: (self.util.iac_get(self.gpib, 
-                                        ("devices/"+self.selected_tec+"/temp_over_state"))), 
-                                    lambda value: self.util.iac_set(self.gpib, 
-                                        ("devices/"+self.selected_tec+"/"), {"temp_over_state": value})),
-                    "set": self.gpib_set,
-                    "info": self.gpib_info,
-                    "temp_sweep": self.gpib_temp_sweep
-                })
-
-        self.param_tree = ParameterTree({
-            "device": self.pico_param_tree,
-            "gpib": self.gpib_tree
-        })
-
-    def verify_gpib(self):
-        """verify that GPIB drivers are installed, USB Device is connected,
-        Adapter is loaded and valid K2510 device is connected"""
+            self.gpib = None
         try:
-            return (
-                bool(self.gpib) and
-                bool(self.gpib.gpibmanager.driver_available) and
-                bool(self.gpib.gpibmanager.device_available) and
-                bool(self.tec_devices)
-            )
-        except AttributeError:
-            return False
+            if self.gpib:
+                devices = self.util.iac_get(self.gpib, "devices")
+                self.gpib_config.tec_devices = [name for name, info in devices.items()
+                    if info.get("type") == "K2510"]
 
-    def get_dc_value(self, obj, chan_name, attr_name):
-        """Retrive values for the live-view settings."""
-        try:
-            channel_dc = getattr(obj, chan_name)
-            return getattr(channel_dc, attr_name, None)
-        except AttributeError:
-            return None
+                self.gpib_config.avail = self.ctrl_util.verify_gpib_avail()
 
-    def set_dc_value(self, obj, attr_name, value):
-        """Change values for the live-view settings."""
-        # Check if PHA needs to be reset
-        if (
-            (attr_name == "num_bins")
-            or (attr_name == "lower_range")
-            or (attr_name == "upper_range")
-        ):
-            self.analysis.clear_pha = True
+                if self.gpib_config.tec_devices:
+                    self.gpib_config.selected_tec = self.gpib_config.tec_devices[0]
 
-        # Ensure a negative value has not been entered
-        if attr_name == "pre_trig_samples" or attr_name == "post_trig_samples" or (
-            attr_name == "auto-trigger_ms") or attr_name == "delay" or (
-                attr_name == "capture_delay"):
-            if value < 0:
-                value = value * (-1)
-
-        # Check setting validity
-        # Check upper range is not lower than lower range, and vice versa
-        if attr_name == "upper_range":
-            if value < self.dev_conf.pha.lower_range:
-                value = self.dev_conf.pha.lower_range + 1
-
-        if attr_name == "lower_range":
-        #     if value < 0:
-        #         value = value * (-1)
-
-
-            if value > self.dev_conf.pha.upper_range:
-                value = self.dev_conf.pha.upper_range - 1
-
-        if attr_name == "num_bins" or attr_name == "n_captures" or attr_name == "repeat_amount":
-            if value < 1:
-                value = 1
-
-        if attr_name == "capture_time":
-            if value <= 0:
-                value = value * (-1)
-
-        setattr(obj, attr_name, value)
-
-    def set_dc_chan_value(self, obj, chan_name, attr_name, value):
-        """Change values for the individual channel settings."""
-        # Ensure the user does not input a negative offset
-        # if attr_name == "offset":
-        #     if value < 0:
-        #         value = value * (-1)
-
-        # Check setting validity
-        # Check if channel selected for LV is actually active
-        if attr_name == "live_view":
-            try:
-                channel_dc = getattr(obj, chan_name)
-                if getattr(channel_dc, "active", None):
-                    setattr(channel_dc, attr_name, value)
-            except AttributeError:
-                pass
-
-        # When channels turn off, ensure they are not LV/PHA active
-        elif (attr_name == "active") and (not value):
-            try:
-                channel_dc = getattr(obj, chan_name)
-                setattr(channel_dc, "live_view", value)
-                setattr(channel_dc, 'pha_active', value)
-                setattr(channel_dc, attr_name, value)
-            except AttributeError:
-                pass
-        else:
-            try:
-                channel_dc = getattr(obj, chan_name)
-                setattr(channel_dc, attr_name, value)
-            except AttributeError:
-                pass
-        # reset pha if pha params change
-        if (
-            (attr_name == "num_bins")
-            or (attr_name == "lower_range")
-            or (attr_name == "upper_range")
-        ):
-            self.analysis.clear_pha = True
-            if attr_name == "num_bins":
-                self.buffer_manager.reset_pha()
-
-    def verify_settings(self):
-        """Verify all picoscope settings, sets status of individual groups of settings."""
-        # Create list of Boolean values as to whether a channel is active or not
-        active = [
-            self.dev_conf.channel_a.active,
-            self.dev_conf.channel_b.active,
-            self.dev_conf.channel_c.active,
-            self.dev_conf.channel_d.active,
-        ]
-
-        # Use functions to verify all of the chosen settings
-        self.pico_status.pico_setup_verify = self.util.verify_mode_settings(
-            active, self.dev_conf.mode
-        )
-        for chan in self.channels:
-            chan.verified = self.util.verify_channel_settings(chan.offset)
-        self.pico_status.channel_setup_verify = self.util.set_channel_verify_flag(
-            self.channels
-        )
-        self.pico_status.channel_trigger_verify = self.util.verify_trigger(
-            self.channels, self.dev_conf.trigger
-        )
-        self.pico_status.capture_settings_verify = self.util.verify_capture(
-            self.dev_conf.capture
-        )
-        self.pico_status.flags.verify_all = self.set_verify_flag()
-
-    def set_verify_flag(self):
-        """Check if PicoScope settings are verified."""
-        status_list = [
-            self.pico_status.pico_setup_verify,
-            self.pico_status.channel_setup_verify,
-            self.pico_status.channel_trigger_verify,
-            self.pico_status.capture_settings_verify,
-        ]
-
-        # Check if there are any issues with the statuses above
-        for status in status_list:
-            if status != 0:
-                return False
-        return True
-
-    def set_capture_run_limits(self):
-        """Set the value for maximum amount of captures that can fit into the picoscope memory."""
-        # Calculate the amount of samples in a capture
-        capture_samples = (
-            self.dev_conf.capture.pre_trig_samples
-            + self.dev_conf.capture.post_trig_samples
-        )
-        # Calculate the maximum amount of captures depending on settings
-        self.dev_conf.capture_run.caps_max = math.floor(
-            self.util.max_samples(self.dev_conf.mode.resolution) / capture_samples
-        )
-
-        if len(self.buffer_manager.active_channels) > 1:
-            self.dev_conf.capture_run.caps_max /= len(self.buffer_manager.active_channels)
-
-        self.dev_conf.capture_run.caps_remaining = self.dev_conf.capture.n_captures
-
-    def set_capture_run_length(self):
-        """Sets the captures to be completed in each "run" based on the maximum allowed captures, and the amount of captures left to be collected"""
-
-        if self.dev_conf.capture_run.caps_remaining <= self.dev_conf.capture_run.caps_max:
-            self.dev_conf.capture_run.caps_in_run = self.dev_conf.capture_run.caps_remaining
-        else:
-            self.dev_conf.capture_run.caps_in_run = self.dev_conf.capture_run.caps_max
-
-    def calc_samp_time(self):
-        """Calculate the sample interval based on the resolution and timebase."""
-        if self.dev_conf.mode.resolution == 0:
-            if (self.dev_conf.mode.timebase) >= 0 and (
-                self.dev_conf.mode.timebase <= 2
-            ):
-                self.dev_conf.mode.samp_time = math.pow(
-                    2, self.dev_conf.mode.timebase
-                ) / (1000000000)
-            else:
-                self.dev_conf.mode.samp_time = (self.dev_conf.mode.timebase - 2) / (
-                    125000000
-                )
-        elif self.dev_conf.mode.resolution == 1:
-            if (self.dev_conf.mode.timebase) >= 1 and (
-                self.dev_conf.mode.timebase <= 3
-            ):
-                self.dev_conf.mode.samp_time = math.pow(
-                    2, self.dev_conf.mode.timebase - 1
-                ) / (500000000)
-            else:
-                self.dev_conf.mode.samp_time = (self.dev_conf.mode.timebase - 3) / (
-                    62500000
-                )
+            # instaniate the parametertree builder classes
+            pico_tree_builder = PicoTreeBuilder(self)
+            gpib_tree_builder = GPIBTreeBuilder(self)
+            
+            # Build both device and GPIB trees
+            device_tree = pico_tree_builder.build_device_tree()
+            gpib_tree = gpib_tree_builder.create_gpib_tree()
+            
+            # Create the complete parameter tree
+            self.param_tree = ParameterTree({
+                "device": device_tree,
+                "gpib": gpib_tree
+            })
+        except Exception as e:
+            logging.error(e)
 
     def run_capture(self):
         """Tell the picoscope to collect and return data."""
 
         # if a single-shot temperature has been set, wait for tec to stablise
-        if self.pico_status.flags.temp_set and self.gpib_control:
+        if self.pico_status.flags.temp_set and self.gpib_config.control_enabled:
             self.pico_status.flags.system_state = "Waiting for TEC to stabilise"
             while self.pico_status.flags.temp_set and not self.pico_status.flags.abort_cap:
                 time.sleep(0.1)
@@ -716,10 +110,10 @@ class PicoController:
         if self.pico_status.flags.temp_reached:
             self.pico_status.flags.temp_reached = False
             
-        self.calc_samp_time()
+        self.ctrl_util.calc_samp_time()
 
         if self.pico_status.flags.verify_all:
-            self.check_res()
+            self.ctrl_util.check_res()
 
             # Check if user has requested a capture
             if self.pico_status.flags.user_capture:
@@ -742,17 +136,17 @@ class PicoController:
                         # reset abort flag if its been set by TB capture
                         self.pico_status.flags.abort_cap = False
                         self.buffer_manager.reset_pha()
-                        self.current_capture = capture_run
+                        self.dev_conf.capture_run.current_capture = capture_run
 
-                        logging.debug(f"current capture repeat: {self.current_capture}")
+                        logging.debug(f"current capture repeat: {self.dev_conf.capture_run.current_capture}")
 
                         # Complete a capture run, based on capture number
                         if not self.dev_conf.capture.capture_type:
                             if not self.pico_status.flags.abort_cap:
                                 # TEC-sweep
-                                if (self.dev_conf.temp_sweep.active and
-                                    self.gpib_control and
-                                    self.gpib_avail):
+                                if (self.gpib_config.active and
+                                    self.gpib_config.control_enabled and
+                                    self.gpib_config.avail):
                                     self.run_temperature_sweep()
                                 else:
                                     self.user_capture(True)
@@ -762,9 +156,9 @@ class PicoController:
                             # Complete a capture run, for a pre-determined amount of time
                             if not self.pico_status.flags.abort_cap:
                                 # TEC-sweep
-                                if (self.dev_conf.temp_sweep.active and
-                                    self.gpib_control and
-                                    self.gpib_avail):
+                                if (self.gpib_config.active and
+                                    self.gpib_config.control_enabled and
+                                    self.gpib_config.avail):
                                     self.run_temperature_sweep()
                                 else:
                                     self.tb_capture()
@@ -784,9 +178,7 @@ class PicoController:
                         if (capture_run + 1) == cap_loop:
                             self.current_time = 0
 
-                    self.current_capture = 0
-                    # deprecated? remove?
-                    #self.file_writer.capture_number = 1
+                    self.dev_conf.capture_run.current_capture = 0
                     self.pico_status.flags.user_capture = False
                     self.pico_status.flags.abort_cap = False
                     self.dev_conf.file.repeat_suffix = None
@@ -810,13 +202,6 @@ class PicoController:
         ):
             self.pico_status.flags.system_state = "Connected to PicoScope, Idle"
 
-    def check_res(self):
-        """Detect if the device resolution has been changed, if so apply to picoscope."""
-        if self.pico_status.flags.res_changed:
-            if self.pico_status.open_unit == 0:
-                self.pico.stop_scope()
-            self.pico_status.flags.res_changed = False
-
     def user_capture(self, save_file):
         """Run the appropriate steps for a set of captures."""
         # Identify whether the capture is a user capture or just for LV
@@ -825,19 +210,19 @@ class PicoController:
         else:
             captures = 2
             
-        self.set_capture_run_limits()
+        self.ctrl_util.set_capture_run_limits()
         
         #set caps_remaining for liveview mode
         if not save_file:
             self.dev_conf.capture_run.caps_remaining = 2
             
-        self.set_capture_run_length()
+        self.ctrl_util.set_capture_run_length()
     
         # checks run_setup completes successfully, calls it with captures if save_file is not true
         if self.pico.run_setup() if save_file else self.pico.run_setup(captures):
             while self.dev_conf.capture_run.caps_comp < captures:
                 if not self.pico_status.flags.abort_cap:
-                    self.set_capture_run_length()
+                    self.ctrl_util.set_capture_run_length()
                     self.capture_run()
                     self.dev_conf.capture_run.caps_remaining -= (
                         self.dev_conf.capture_run.caps_in_run
@@ -869,7 +254,7 @@ class PicoController:
         """
         """
         # validate this method of calculating max captures!
-        self.set_capture_run_limits()
+        self.ctrl_util.set_capture_run_limits()
         self.dev_conf.capture_run.caps_in_run = int(self.dev_conf.capture_run.caps_max/2)
         self.pico.run_time_based_capture(
             self.dev_conf.capture.capture_time
@@ -877,128 +262,29 @@ class PicoController:
         self.file_writer.write_hdf5(write_accumulated=True)
         self.buffer_manager.clear_arrays()
         self.pico_status.flags.abort_cap = False
-
-    def wait_for_tec(self, target: float, tol: float):
-        """
-        Simple temperature stability check with basic logging.
-        Waits for temperature to remain stable within tolerance for specified time.
-        """
-        from collections import deque
-        
-        sweep_config = self.dev_conf.temp_sweep
-
-        # Simple time-based calculation
-        stability_readings = int(sweep_config.stability_time / sweep_config.poll_s)
-        errors = deque(maxlen=stability_readings)
-        
-        start_time = time.time()
-        logging.info(f"[TEC] Waiting for {target:.2f}°C ± {tol:.2f}°C")  
-        reading_count = 0
-        
-        while not self.pico_status.flags.abort_cap:
-            # Get current temperature
-            meas = self.util.iac_get(self.gpib, f"devices/{self.selected_tec}/info/tec_temp_meas")
-            if meas is None:
-                logging.error("[TEC] Failed to read temperature")
-                break
-                
-            # Calculate error and add to rolling window
-            error = meas - target
-            errors.append(error)
-            reading_count += 1
-            
-            if reading_count % 10 == 0:
-                elapsed = time.time() - start_time
-                mean_error = sum(abs(e) for e in errors) / len(errors)
-                logging.debug(f"[TEC] Current: {meas:.2f}°C, Avg error: ±{mean_error:.3f}°C "
-                            f"(readings: {len(errors)}/{stability_readings}, elapsed: {elapsed:.1f}s)")
-            
-            # Check for stability once we have enough readings
-            if len(errors) >= stability_readings:
-                mean_error = sum(abs(e) for e in errors) / len(errors)
-                
-                if mean_error <= tol:
-                    elapsed_time = time.time() - start_time
-                    actual_stability_time = len(errors) * sweep_config.poll_s
-                    logging.info(f"[TEC] Temperature stable at {meas:.2f}°C "
-                                f"(±{mean_error:.3f}°C over {actual_stability_time:.1f}s, "
-                                f"total time: {elapsed_time:.1f}s)")
-                    return
-            time.sleep(sweep_config.poll_s)
-        
-    def _temp_range(self, start, end, step):
-        """
-        Return a list of temperatures in equal increments, direction is 
-        inferred from start   end
-        """
-        if step == 0 or start == end:
-            return [start]
-
-        step      = abs(step) # ignore sign the user gave
-        direction = 1 if end >= start else -1
-        temps     = [start]
-        current   = start
-
-        while True:
-            next_val = current + direction * step
-            # Would the next step cross the end value?
-            if (direction == 1 and next_val >= end) or \
-            (direction == -1 and next_val <= end):
-                break
-            temps.append(next_val)
-            current = next_val
-
-        if temps[-1] != end:
-            temps.append(end)
-
-        logging.debug(f"returning temps: {temps}")
-        return temps
-    
-    def _temp_suffix(self, T: float) -> str:
-        """
-        Return a filename-safe suffix like '_25-0c' or '_-5-0c'
-        """
-        s = f"{T:.1f}".replace(".", "-")
-        return f"_{s}c"
-
-    def _clean_base_fname(self) -> str:
-        """
-        Return the users current file name without .hdf5 and other suffix's    
-        """
-        name = self.dev_conf.file.file_name
-        if name.endswith(".hdf5"):
-            name = name[:-5]
-
-        # strip “…_<digits>”  (repeat suffix)
-        name = re.sub(r"_\d+$", "", name)
-
-        # strip “…_<±number>[ - number]c”  (temperature suffix)
-        name = re.sub(r"_-?\d+(?:-\d+)?c$", "", name)
-
-        return name.rstrip("_") 
-    
+   
     def run_temperature_sweep(self):
         """
         Iterate over every temperature in the listeven in SIM mode
         and acquire data.  A guard flag prevents any single capture from
         setting `abort_cap` and killing the rest of the sweep.
         """
-        sweep = self.dev_conf.temp_sweep
+        sweep = self.gpib_config
         if not sweep.active:
             return
 
-        temps    = self._temp_range(sweep.t_start, sweep.t_end, sweep.t_step)
-        self.dev_conf.temp_sweep.sweep_points = len(temps)
+        temps    = self.ctrl_util.temp_range(sweep.t_start, sweep.t_end, sweep.t_step)
+        self.gpib_config.sweep_points = len(temps)
         logging.info(f"[TEC-sweep] Set-points: {temps}")
 
-        base_fname = self._clean_base_fname()
+        base_fname = self.ctrl_util.clean_base_fname()
 
         # local flag so abort in one capture doesn’t cancel the sweep
         sweep_abort = False
 
         for idx, T in enumerate(temps):
-            self.dev_conf.temp_sweep.sweep_index  = idx
-            logging.debug(f"Current temp sweep: {self.dev_conf.temp_sweep.sweep_index}")
+            self.gpib_config.sweep_index  = idx
+            logging.debug(f"Current temp sweep: {self.gpib_config.sweep_index}")
             if sweep_abort:
                 break
 
@@ -1008,7 +294,7 @@ class PicoController:
             else:
                 self.util.iac_set(
                     self.gpib,
-                    f"devices/{self.selected_tec}/set/temp_set",
+                    f"devices/{self.gpib_config.selected_tec}/set/temp_set",
                     float(T)
                 )
             self.buffer_manager.temp_set_last = T
@@ -1019,15 +305,15 @@ class PicoController:
                 time.sleep(sweep.poll_s)
                 self.buffer_manager.temp_meas_last = T
             else:
-                self.wait_for_tec(T, sweep.tol)
+                self.ctrl_util.wait_for_tec(T, sweep.tol)
                 self.buffer_manager.temp_meas_last = self.util.iac_get(
                     self.gpib,
-                    f"devices/{self.selected_tec}/info/tec_temp_meas")
+                    f"devices/{self.gpib_config.selected_tec}/info/tec_temp_meas")
                 
             # self.dev_conf.file.file_name = (
-            #     base_fname + self._temp_suffix(T)
+            #     base_fname + self.ctrl_util.temp_suffix(T)
             # )
-            self.dev_conf.file.temp_suffix = str(self._temp_suffix(T))
+            self.dev_conf.file.temp_suffix = str(self.ctrl_util.temp_suffix(T))
 
             # reset abort flag before each capture; if *this* capture sets it,
             # finish the file and then abandon the remaining temps
@@ -1047,15 +333,15 @@ class PicoController:
     
         # clear temperature suffix and reset sweep index to 0
         self.dev_conf.file.temp_suffix = None
-        self.dev_conf.temp_sweep.sweep_points = 0
-        self.dev_conf.temp_sweep.sweep_index  = 0 
+        self.gpib_config.sweep_points = 0
+        self.gpib_config.sweep_index  = 0 
 
     def set_temp_single_shot(self, _=None):
         """
         Called when the user clicks /gpib/set/set_temp.
         Reads temp_target and kicks off the background wait.
         """
-        T = self.dev_conf.temp_sweep.temp_target
+        T = self.gpib_config.temp_target
         self.pico_status.flags.temp_set     = True
         self.pico_status.flags.temp_reached = False
         self.pico_status.flags.system_state = f"Setting TEC {T:.2f} °C"
@@ -1070,12 +356,21 @@ class PicoController:
         c) flip flags & update status
         """
         # a)
-        self.util.iac_set(self.gpib,
-                        f"devices/{self.selected_tec}/set/temp_set",
-                        float(T))
+        try:
+            logging.debug(f"Temp : {T}")
+            self.util.iac_set(self.gpib,
+                            f"devices/{self.gpib_config.selected_tec}/set",
+                            {"temp_set": float(T)})
+            val = self.util.iac_get(
+                self.gpib,
+                f'devices/{self.gpib_config.selected_tec}/set/temp_set'
+            )
+            logging.debug(f"iac_get returned: {val}")
+        except Exception as e:
+            logging.error(e)
         # b)
-        self.wait_for_tec(T, self.dev_conf.temp_sweep.tol)
-        # c)
+        self.ctrl_util.wait_for_tec(T, self.gpib_config.tol)
+            # c)
         self.pico_status.flags.temp_set     = False
         self.pico_status.flags.temp_reached = True
 
@@ -1086,6 +381,7 @@ class PicoController:
     def update_loop(self):
         """Execute thread, responsible for calling the run_capture function at timed intervals."""
         while self.update_loop_active:
+            #logging.debug(f"temp target : {(self.gpib_config.temp_target)}")
             self.run_capture()
             time.sleep(0.2)
 
@@ -1110,7 +406,7 @@ class PicoController:
             self.param_tree.set(path, data)
         except ParameterTreeError as e:
             raise PicoControllerError(e)
-        self.verify_settings()
+        self.ctrl_util.verify_settings()
 
 class PicoControllerError(Exception):
     pass
